pyproject.toml
[build-system]
requires = ['setuptools>=42', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'clalign'
version = '1.0'
description = ''
readme = ''
requires-python = '==3.12.7'
authors = [
    {name = 'yourh', email = 'yourh@nankai.com'}
]
license = {text = 'Custom'}
classifiers = []
dependencies = [
    'numpy>=2.0.2',
    'scipy>=1.14.1',
    'biopython>=1.84',
    'torch==2.4.1+cu124',
    'transformers>=4.41.2',
    'sentencepiece>=0.2.0',
    'protobuf>=5.28.3',
    'peft>=0.13.0',
    'numba>=0.60.0',
    'click>=8.1.7',
    'logzero>=1.7.0',
    'tqdm>=4.66.5'
]

[project.urls]
homepage = 'https://github.com/yourh/CLAlign'

[project.scripts]
clalign = 'clalign.main:main'

[tool.setuptools.package-data]
clalign = ['CLAlignT5/**/*']

[tool.setuptools.packages.find]
include = ['clalign']
